+++ gambler/source/Fold/Nonempty/Conversion.hs
@@ -5,9 +5,12 @@ import Fold.Nonempty.Type
 
 import Fold.Effectful.Type (EffectfulFold)
 import Fold.Pure.Type (Fold (Fold))
+import Fold.ShortcutNonempty.Type (ShortcutNonemptyFold (ShortcutNonemptyFold))
+import Strict (shortcut)
 
 import qualified Fold.Pure.Type as Fold
 import qualified Fold.Pure.Conversion as Fold.Conversion
+import qualified Fold.ShortcutNonempty.Type as Shortcut
 
 import Data.Functor.Identity (Identity)
 
@@ -21,3 +24,9 @@ fold Fold{ Fold.initial, Fold.step, Fold.extract } =
 one input -}
 effectfulFold :: EffectfulFold Identity a b -> NonemptyFold a b
 effectfulFold x = fold (Fold.Conversion.effectfulFold x)
+
+shortcutNonemptyFold :: ShortcutNonemptyFold a b -> NonemptyFold a b
+shortcutNonemptyFold ShortcutNonemptyFold{
+        Shortcut.step, Shortcut.initial, Shortcut.extract } =
+    NonemptyFold{ initial = \a -> shortcut (initial a),
+        step = \x a -> shortcut (step x a), extract }